+++ Userland/Applications/VideoPlayer/main.cpp
@@ -43,8 +43,8 @@ int main(int argc, char** argv)
 
             auto const& frame = block.frame(0);
             dbgln("Reading frame 0 from block @ {}", block.timestamp());
-            vp9_decoder.parse_frame(frame);
-            vp9_decoder.dump_info();
+            vp9_decoder.decode_frame(frame);
+            vp9_decoder.dump_frame_info();
         }
     }
