+++ Userland/Libraries/LibVideo/CMakeLists.txt
@@ -5,6 +5,7 @@ set(SOURCES
     VP9/Decoder.cpp
     VP9/Enums.h
     VP9/LookupTables.h
+    VP9/Parser.cpp
     VP9/ProbabilityTables.cpp
     VP9/Symbols.h
     VP9/SyntaxElementCounter.cpp
